+++ src/pages/heewonKang/Main/Main.scss
@@ -69,7 +69,7 @@ main {
 
             /* 스토리 화살표 2개 */
             .arrows {
-              @include pos-center(absolute);
+              position: absolute;
               display: flex;
               justify-content: space-between;
               align-items: center;
